+++ docs/NEWS.md
@@ -15,6 +15,10 @@ respective loader.
 - Added `engine_check` to check whether npm and/or yarn is 
 currently and can be found: necessary to run anything with 
 packer.
+- Simplify check for packer, no longer requires `dev` directory
+to be present.
+- Added `put_test` to add a testthat check to ensure files are
+minified.
 
 ## packer 0.1.1