+++ docs/NEWS.md
@@ -1,3 +1,4 @@
+# Changelog
 ## packer 0.1.2
 
 - Few typo in documentation.